Abstract AIController, switch to control codes
