+++ core/src/com/fallenflame/game/LevelModel.java
@@ -2,7 +2,7 @@ package com.fallenflame.game;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.math.*;
-import com.badlogic.gdx.physics.box2d.PolygonShape;
+import com.fallenflame.game.enemies.EnemyModel;
 import com.fallenflame.game.physics.obstacle.BoxObstacle;
 import com.fallenflame.game.physics.obstacle.WheelObstacle;
